Rename can_logger/player.py to canlogger/player
Move can_logger.py to module can.io.logger
Move can_player.py to module can.io.player
Make entry points canplayer and canlogger
Move Logger and LogReader classes to logger.py and player.py
