+++ setup.py
@@ -23,12 +23,9 @@ setup(
         "": ["CONTRIBUTORS.txt", "LICENSE.txt"],
         "doc": ["*.*"]
     },
-
-    scripts=[
-        "./bin/can_logger.py",
-        "./bin/can_player.py"
-    ],
     entry_points={"console_scripts": [
+        "canlogger = can.io.logger:main",
+        "canplayer = can.io.player:main",
         "canserver = can.interfaces.remote.__main__:main"
     ]},
